requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
langchain-openai>=0.0.1
python-dotenv>=1.0.0
faiss-cpu>=1.7.0
pypdf>=3.0.0
tiktoken>=0.3.0